[Refactor] search 패키지 변경
