+++ app/src/main/java/com/doctor/yumyum/presentation/ui/researchlist/ResearchListActivity.kt
@@ -19,8 +19,8 @@ import com.doctor.yumyum.databinding.DialogSelectSortBinding
 import com.doctor.yumyum.presentation.adapter.TasteTagAdapter
 import com.doctor.yumyum.presentation.ui.filter.FilterActivity
 import com.doctor.yumyum.presentation.ui.recipedetail.RecipeDetailActivity
-import com.doctor.yumyum.presentation.ui.searchhashtag.SearchHashtagActivity
-import com.doctor.yumyum.presentation.ui.searchtaste.SearchTasteActivity
+import com.doctor.yumyum.presentation.ui.search.hashtag.SearchHashtagActivity
+import com.doctor.yumyum.presentation.ui.search.taste.SearchTasteActivity
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import kotlinx.coroutines.launch
 import java.util.*
