+++ app/src/main/res/layout/activity_search_taste.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="viewModel"
-            type="com.doctor.yumyum.presentation.ui.searchtaste.SearchTasteViewModel" />
+            type="com.doctor.yumyum.presentation.ui.search.taste.SearchTasteViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
@@ -13,7 +13,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.searchtaste.SearchTasteActivity">
+        tools:context=".presentation.ui.search.taste.SearchTasteActivity">
 
         <include
             android:id="@+id/search_taste_appbar"
